THPT2024_Analyzer/requirements.txt
django
djangorestframework
psycopg2-binary
confluent_kafka
matplotlib
numpy
